LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-text/white-space/seg-break-transformation-004-expected.txt


FAIL linebreak only ₩２４ assert_equals: expected 72 but got 79
FAIL spaces linebreak ₩２４ assert_equals: expected 72 but got 79
FAIL linebreak spaces ₩２４ assert_equals: expected 72 but got 79
FAIL spaces linebreak spaces ₩２４ assert_equals: expected 72 but got 79
FAIL multiple linebreaks ₩２４ assert_equals: expected 72 but got 79
FAIL multiple linebreaks + spaces ₩２４ assert_equals: expected 72 but got 79
FAIL linebreak only ２４₩ assert_equals: expected 72 but got 79
FAIL spaces linebreak ２４₩ assert_equals: expected 72 but got 79
FAIL linebreak spaces ２４₩ assert_equals: expected 72 but got 79
FAIL spaces linebreak spaces ２４₩ assert_equals: expected 72 but got 79
FAIL multiple linebreaks ２４₩ assert_equals: expected 72 but got 79
FAIL multiple linebreaks + spaces ２４₩ assert_equals: expected 72 but got 79
₩ ２４
₩ ２４
₩ ２４
₩ ２４
₩ ２４
₩ ２４
₩２４
２４ ₩
２４ ₩
２４ ₩
２４ ₩
２４ ₩
２４ ₩
２４₩
